Move runtime boundary DTOs to shared contracts
