+++ implementations/python/packages/aces_backend_protocols/protocols.py
@@ -2,21 +2,31 @@
 
 from __future__ import annotations
 
-from typing import Any, Protocol
+from typing import Protocol
+
+from aces_contracts.diagnostics import Diagnostic
+from aces_contracts.participant_episode import (
+    ParticipantEpisodeInitializeRequest,
+    ParticipantEpisodeResetRequest,
+    ParticipantEpisodeRestartRequest,
+    ParticipantEpisodeTerminateRequest,
+)
+from aces_contracts.planning import EvaluationPlan, OrchestrationPlan, ProvisioningPlan
+from aces_contracts.runtime_state import ApplyResult, RuntimeSnapshot
 
 
 class Provisioner(Protocol):
     """Applies provisioning plans to the target environment."""
 
-    def validate(self, plan: Any) -> list[Any]:
+    def validate(self, plan: ProvisioningPlan) -> list[Diagnostic]:
         """Return planner/runtime diagnostics for an apply attempt."""
         ...
 
     def apply(
         self,
-        plan: Any,
-        snapshot: Any,
-    ) -> Any:
+        plan: ProvisioningPlan,
+        snapshot: RuntimeSnapshot,
+    ) -> ApplyResult:
         """Apply provisioning reconciliation operations."""
         ...
 
@@ -26,25 +36,25 @@ class Orchestrator(Protocol):
 
     def start(
         self,
-        plan: Any,
-        snapshot: Any,
-    ) -> Any:
+        plan: OrchestrationPlan,
+        snapshot: RuntimeSnapshot,
+    ) -> ApplyResult:
         """Start or refresh orchestration state."""
         ...
 
-    def status(self) -> dict[str, Any]:
+    def status(self) -> dict[str, object]:
         """Return current orchestration status."""
         ...
 
-    def results(self) -> dict[str, dict[str, Any]]:
+    def results(self) -> dict[str, dict[str, object]]:
         """Return most recent workflow execution state envelope."""
         ...
 
-    def history(self) -> dict[str, list[dict[str, Any]]]:
+    def history(self) -> dict[str, list[dict[str, object]]]:
         """Return workflow execution history events."""
         ...
 
-    def stop(self, snapshot: Any) -> Any:
+    def stop(self, snapshot: RuntimeSnapshot) -> ApplyResult:
         """Stop orchestration and clear orchestration state."""
         ...
 
@@ -54,25 +64,25 @@ class Evaluator(Protocol):
 
     def start(
         self,
-        plan: Any,
-        snapshot: Any,
-    ) -> Any:
+        plan: EvaluationPlan,
+        snapshot: RuntimeSnapshot,
+    ) -> ApplyResult:
         """Start or refresh evaluation state."""
         ...
 
-    def status(self) -> dict[str, Any]:
+    def status(self) -> dict[str, object]:
         """Return current evaluator status."""
         ...
 
-    def results(self) -> dict[str, dict[str, Any]]:
+    def results(self) -> dict[str, dict[str, object]]:
         """Return most recent evaluation results."""
         ...
 
-    def history(self) -> dict[str, list[dict[str, Any]]]:
+    def history(self) -> dict[str, list[dict[str, object]]]:
         """Return evaluation history events."""
         ...
 
-    def stop(self, snapshot: Any) -> Any:
+    def stop(self, snapshot: RuntimeSnapshot) -> ApplyResult:
         """Stop evaluation and clear evaluation state."""
         ...
 
@@ -92,17 +102,17 @@ class ParticipantRuntime(Protocol):
 
     def initialize(
         self,
-        request: Any,
-        snapshot: Any,
-    ) -> Any:
+        request: ParticipantEpisodeInitializeRequest,
+        snapshot: RuntimeSnapshot,
+    ) -> ApplyResult:
         """Create the first episode for a participant (sequence_number=0)."""
         ...
 
     def reset(
         self,
-        request: Any,
-        snapshot: Any,
-    ) -> Any:
+        request: ParticipantEpisodeResetRequest,
+        snapshot: RuntimeSnapshot,
+    ) -> ApplyResult:
         """Start a new episode instance from a non-terminal predecessor.
 
         Must allocate a new ``episode_id`` and increment
@@ -113,9 +123,9 @@ class ParticipantRuntime(Protocol):
 
     def restart(
         self,
-        request: Any,
-        snapshot: Any,
-    ) -> Any:
+        request: ParticipantEpisodeRestartRequest,
+        snapshot: RuntimeSnapshot,
+    ) -> ApplyResult:
         """Start a new episode instance from a terminated predecessor.
 
         Must allocate a new ``episode_id`` and increment
@@ -126,20 +136,20 @@ class ParticipantRuntime(Protocol):
 
     def terminate(
         self,
-        request: Any,
-        snapshot: Any,
-    ) -> Any:
+        request: ParticipantEpisodeTerminateRequest,
+        snapshot: RuntimeSnapshot,
+    ) -> ApplyResult:
         """Drive the current episode to ``TERMINATED`` with the given reason."""
         ...
 
-    def status(self) -> dict[str, Any]:
+    def status(self) -> dict[str, object]:
         """Return current participant runtime status."""
         ...
 
-    def results(self) -> dict[str, dict[str, Any]]:
+    def results(self) -> dict[str, dict[str, object]]:
         """Return the most recent participant episode result envelopes."""
         ...
 
-    def history(self) -> dict[str, list[dict[str, Any]]]:
+    def history(self) -> dict[str, list[dict[str, object]]]:
         """Return participant episode history events."""
         ...
