+++ implementations/python/packages/aces_backend_stubs/stubs.py
@@ -19,15 +19,9 @@ from aces_backend_protocols.capabilities import (
     WorkflowStatePredicateFeature,
 )
 from aces_contracts.apparatus import ConceptBinding, RealizationSupportDeclaration
+from aces_contracts.diagnostics import Diagnostic
 from aces_contracts.manifest_authority import BACKEND_SUPPORTED_CONTRACT_IDS
-from aces_contracts.vocabulary import RealizationSupportMode
-from aces_processor.models import (
-    EVALUATION_STATE_SCHEMA_VERSION,
-    ApplyResult,
-    ChangeAction,
-    Diagnostic,
-    EvaluationPlan,
-    OrchestrationPlan,
+from aces_contracts.participant_episode import (
     ParticipantEpisodeControlAction,
     ParticipantEpisodeExecutionState,
     ParticipantEpisodeHistoryEvent,
@@ -38,11 +32,11 @@ from aces_processor.models import (
     ParticipantEpisodeStatus,
     ParticipantEpisodeTerminalReason,
     ParticipantEpisodeTerminateRequest,
-    ProvisioningPlan,
-    RuntimeDomain,
-    RuntimeSnapshot,
-    SnapshotEntry,
 )
+from aces_contracts.planning import ChangeAction, EvaluationPlan, OrchestrationPlan, ProvisioningPlan, RuntimeDomain
+from aces_contracts.runtime_state import ApplyResult, RuntimeSnapshot, SnapshotEntry
+from aces_contracts.versions import EVALUATION_STATE_SCHEMA_VERSION
+from aces_contracts.vocabulary import RealizationSupportMode
 from aces_runtime.registry import RuntimeTarget, RuntimeTargetComponents
 
 REFERENCE_BACKEND_SUPPORTED_CONTRACT_VERSIONS = BACKEND_SUPPORTED_CONTRACT_IDS
