+++ implementations/python/packages/aces_conformance/conformance.py
@@ -35,25 +35,24 @@ from aces_contracts.contracts import (
     WorkflowHistoryEventModel,
     schema_bundle,
 )
-from aces_processor.compiler import compile_runtime_model
-from aces_processor.models import (
-    Diagnostic,
-    EvaluationExecutionState,
-    ParticipantActionContractRuntime,
-    ParticipantBehaviorHistoryEvent,
+from aces_contracts.diagnostics import Diagnostic, Severity
+from aces_contracts.evaluation import EvaluationExecutionState
+from aces_contracts.participant_episode import (
     ParticipantEpisodeExecutionState,
     ParticipantEpisodeHistoryEvent,
     ParticipantEpisodeTerminalReason,
+    iter_participant_episode_snapshot_violations,
+)
+from aces_contracts.planning import RuntimeDomain
+from aces_contracts.runtime_state import RuntimeSnapshot, RuntimeSnapshotEnvelope, SnapshotEntry
+from aces_contracts.workflow import WorkflowExecutionState
+from aces_processor.compiler import compile_runtime_model
+from aces_processor.models import (
+    ParticipantActionContractRuntime,
+    ParticipantBehaviorHistoryEvent,
     ParticipantObservationBoundaryRuntime,
-    RuntimeDomain,
-    RuntimeSnapshot,
-    RuntimeSnapshotEnvelope,
-    Severity,
-    SnapshotEntry,
-    WorkflowExecutionState,
     iter_participant_behavior_history_violations,
     iter_participant_behavior_joint_action_violations,
-    iter_participant_episode_snapshot_violations,
 )
 from aces_processor.planner import plan
 from aces_runtime.control_plane import RuntimeControlPlane
