+++ implementations/python/packages/aces_contracts/planning.py
@@ -1,0 +1,116 @@
+"""Shared runtime planning contracts."""
+
+from dataclasses import dataclass, field
+from enum import Enum
+from typing import Any
+
+from aces_contracts.diagnostics import Diagnostic
+
+
+class RuntimeDomain(str, Enum):
+    """Top-level runtime concern."""
+
+    PROVISIONING = "provisioning"
+    ORCHESTRATION = "orchestration"
+    EVALUATION = "evaluation"
+    PARTICIPANT = "participant"
+
+
+class ChangeAction(str, Enum):
+    """Planner reconciliation result for a resource."""
+
+    CREATE = "create"
+    UPDATE = "update"
+    DELETE = "delete"
+    UNCHANGED = "unchanged"
+
+
+@dataclass(frozen=True)
+class PlannedResource:
+    """Normalized resource used by the planner and snapshot."""
+
+    address: str
+    domain: RuntimeDomain
+    resource_type: str
+    payload: dict[str, Any]
+    ordering_dependencies: tuple[str, ...] = ()
+    refresh_dependencies: tuple[str, ...] = ()
+
+
+@dataclass(frozen=True)
+class PlanOperation:
+    """A reconciliation operation for a planned resource."""
+
+    action: ChangeAction
+    address: str
+    resource_type: str
+    payload: dict[str, Any]
+    ordering_dependencies: tuple[str, ...] = ()
+    refresh_dependencies: tuple[str, ...] = ()
+
+
+class ProvisionOp(PlanOperation):
+    """Provisioning reconciliation operation."""
+
+
+class OrchestrationOp(PlanOperation):
+    """Orchestration reconciliation operation."""
+
+
+class EvaluationOp(PlanOperation):
+    """Evaluation reconciliation operation."""
+
+
+@dataclass(frozen=True)
+class ProvisioningPlan:
+    """Provisioning plan over canonical deployment resources."""
+
+    resources: dict[str, PlannedResource] = field(default_factory=dict)
+    operations: list[ProvisionOp] = field(default_factory=list)
+    diagnostics: list[Diagnostic] = field(default_factory=list)
+
+    @property
+    def actionable_operations(self) -> list[ProvisionOp]:
+        return [op for op in self.operations if op.action != ChangeAction.UNCHANGED]
+
+
+@dataclass(frozen=True)
+class OrchestrationPlan:
+    """Resolved orchestration graph and reconciliation actions."""
+
+    resources: dict[str, PlannedResource] = field(default_factory=dict)
+    operations: list[OrchestrationOp] = field(default_factory=list)
+    startup_order: list[str] = field(default_factory=list)
+    diagnostics: list[Diagnostic] = field(default_factory=list)
+
+    @property
+    def actionable_operations(self) -> list[OrchestrationOp]:
+        return [op for op in self.operations if op.action != ChangeAction.UNCHANGED]
+
+
+@dataclass(frozen=True)
+class EvaluationPlan:
+    """Resolved evaluation graph and reconciliation actions."""
+
+    resources: dict[str, PlannedResource] = field(default_factory=dict)
+    operations: list[EvaluationOp] = field(default_factory=list)
+    startup_order: list[str] = field(default_factory=list)
+    diagnostics: list[Diagnostic] = field(default_factory=list)
+
+    @property
+    def actionable_operations(self) -> list[EvaluationOp]:
+        return [op for op in self.operations if op.action != ChangeAction.UNCHANGED]
+
+
+__all__ = (
+    "ChangeAction",
+    "EvaluationOp",
+    "EvaluationPlan",
+    "OrchestrationOp",
+    "OrchestrationPlan",
+    "PlanOperation",
+    "PlannedResource",
+    "ProvisionOp",
+    "ProvisioningPlan",
+    "RuntimeDomain",
+)
